backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pandas==2.2.3
scikit-learn==1.5.2
pydantic==2.9.2
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.8.3
numpy==2.1.3
requests==2.32.3
aiohttp==3.10.10
sentence-transformers==3.0.1
transformers==4.46.0
torch==2.4.1
langchain==0.2.17
langchain-community==0.2.8
langchain-google-genai==0.0.14
httpx==0.28.1
psutil==6.0.0
matplotlib==3.9.2